improve function that determines whether a face is near/far
The previous method of determining the distance (near/far) of a face averages the z
values of each vertex that makes up the face. This method is ok for
triangles that are equalateral (or almost equalateral) but does not work
as well for triangles as they become less and less equalateral. The new
method uses the normal vector of the face to determine its distance.
This method works for all triangle.

- util:
    * new function to calculate the normal vector of a face
    * new function to determine near/far based on normal vector
- render: update to use new function