fix bug: dists now change properly regardless of rotation speed
issue: When rotating the rendered structure too fast, many nodes, edges,
and faces are not switching from 'front' to 'back' or vice versa, resulting in
incorrect rendering order.

This was caused by the change threshold not adapting to the rotation
step.

solution: account for rotation step: for high rotation steps, use the
highest step value as the threshold that determines when a
node/edge/face needs to switch distance